#101: Traits moved into their own module. This way code keeps cleaner.
